Introduce plugin loader class
