+++ nyanyabot/core/nyanyabot.py
@@ -1,13 +1,9 @@
 import html
 import logging
-import operator
-import sys
 import traceback
-from importlib import import_module
 from io import StringIO
 from queue import Queue
 
-from sqlalchemy import select
 from telegram import Bot, ParseMode
 from telegram.error import Unauthorized
 from telegram.ext import Defaults, Updater, JobQueue, CallbackContext
@@ -16,6 +12,7 @@ from telegram.utils.request import Request
 from nyanyabot.core.configuration import Configuration
 from nyanyabot.core.constants import Constants
 from nyanyabot.core.dispatcher import Dispatcher
+from nyanyabot.core.pluginloader import PluginLoader
 from nyanyabot.database.database import Database
 
 
@@ -64,7 +61,6 @@ class NyaNyaBot:
         self.config = config
         self.database = Database(config.database_name, config.database_user, config.database_password,
                                  config.database_host, config.database_port)
-        self.plugins = []  # type: ignore
 
         self.logger = logging.getLogger(__name__)
         self.logger.info("Starting up bot")
@@ -93,67 +89,7 @@ class NyaNyaBot:
         self.updater.dispatcher.job_queue.set_dispatcher(self.updater.dispatcher)  # type: ignore
         self.updater.dispatcher.add_error_handler(self.error_handler, run_async=True)
 
-        self.load_plugins()
-
-    def load_plugins(self):
-        plugins = []
-        group = 0
-
-        # Load enabled plugins from database
-        with self.database.engine.begin() as conn:
-            for row in conn.execute(
-                    select(self.database.tables.bot_plugins.c.name).where(
-                            self.database.tables.bot_plugins.c.enabled == 1
-                    )
-            ):
-                plugins.append(row.name)
-
-        # Setup plugins
-        for plugin_dir in self.config.plugin_dirs:
-            sys.path.append(plugin_dir)
-
-        for num, plugin in enumerate(plugins):
-            self.logger.info("(%s/%s) Loading plugin: %s", num + 1, len(plugins), plugin)
-            try:
-                module = import_module("nyanyabot.plugin." + plugin)
-            except ModuleNotFoundError:
-                try:
-                    module = import_module("plugins." + plugin)
-                except Exception:
-                    self.logger.error("".join(traceback.format_exc()))
-                    continue
-            except Exception:
-                self.logger.error("".join(traceback.format_exc()))
-                continue
-
-            # Setup handlers, jobs and commands
-            if hasattr(module, "plugin"):  # Is Plugin?
-                try:
-                    # Call the "plugin" variable which references the specific plugin class
-                    plugin_module = module.plugin(self)  # type: ignore
-                except Exception:
-                    self.logger.error("".join(traceback.format_exc()))
-                    continue
-
-                for handler in plugin_module.handlers:
-                    handler.name = plugin_module.name
-                    self.updater.dispatcher.add_handler(handler, group=group)
-                    group += 1
-
-                self.plugins.append(plugin_module)
-
-    def setup_commands(self):
-        commands = []
-        for plugin in self.plugins:
-            for command in plugin.commands:
-                commands.append(command)
-
-        self.logger.info("Setting commands...")
-        if len(commands) > 100:
-            self.logger.warning("More than 100 commands detected, skipping")
-        else:
-            commands.sort(key=operator.attrgetter("command"))
-            self.updater.bot.set_my_commands(commands)
+        self.plugin_loader = PluginLoader(self)
 
     def start(self):
         try:
@@ -193,7 +129,7 @@ class NyaNyaBot:
             )
 
         if self.config.set_commands_enabled:
-            self.setup_commands()
+            self.plugin_loader.setup_commands()
 
         self.logger.info("Bot completely loaded")
         self.updater.idle()
